requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
mysql-connector-python>=8.0.26
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
boto3>=1.26.0
accelerate>=0.20.0  # Required for model offloading
safetensors>=0.3.1  # Required for model loading
einops>=0.6.1  # Required by transformers
numpy>=1.24.0  # Required by torch and transformers
tqdm>=4.65.0  # Required for progress bars
requests>=2.31.0  # Required for HTTP requests
